Changed `IsStrict` class to use `TypeFamilies`, not `FunctionalDependencies`.
`IsStrict` is now a single parameter type class using a type family instance.

The main reason I've changed this is to avoid orphan instances in all the modules.

This is a backwards compatability breaking change.
Any instance declarations of `IsStrict` will need to be adjusted,
Also any mention of `IsStrict` in type signatures will need to be changed.